requirements.txt
cli-helpers==2.1.0
click==7.1.2
configobj==5.0.6
Django==2.2.17
django-cors-headers==3.5.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
humanize==2.6.0
pgcli==2.2.0
pgspecial==1.11.10
prompt-toolkit==2.0.10
psycopg2==2.8.6
Pygments==2.7.4
PyJWT==1.7.1
pytz==2020.5
setproctitle==1.1.10
six==1.15.0
sqlparse==0.3.1
tabulate==0.8.7
terminaltables==3.1.0
wcwidth==0.2.5
